Restrucuracion de los archivos para que funcione como en master
