+++ lib/screens/landing/landing.dart
@@ -5,7 +5,7 @@ class Landing extends StatelessWidget {
 
   @override
   Widget build(BuildContext context) {
-    return Scaffold(
+    return Scaffold( backgroundColor: Color(0xffE0DDCF),
       body: SingleChildScrollView(
         child: Column(
           mainAxisAlignment: MainAxisAlignment.start,
