+++ lib/screens/landing/widgets/landingcard.dart
@@ -1,0 +1,40 @@
+import 'package:auto_matic/config/config.dart';
+
+class Landingcard extends StatefulWidget {
+  const Landingcard(
+      {super.key,
+      required this.titulo,
+      required this.texto,
+      required this.imagen});
+  final String titulo;
+  final String texto;
+  final String imagen;
+
+  @override
+  State<Landingcard> createState() => _LandingcardState();
+}
+
+class _LandingcardState extends State<Landingcard> {
+  @override
+  Widget build(BuildContext context) {
+    return Container(
+      height: 100,
+      margin: EdgeInsets.all(10),
+      child: Card.filled(
+        color: const Color(0xff998fc7),
+        child: Column(
+          mainAxisSize: MainAxisSize.min,
+          children: <Widget>[
+            Expanded(
+              child: ListTile(
+                leading: Image.asset(widget.imagen),
+                title: Text(widget.titulo),
+                subtitle: Text(widget.texto),
+              ),
+            )
+          ],
+        ),
+      ),
+    );
+  }
+}
