--- lib/widgets/widgets.dart
@@ -1,1 +1,0 @@
-export 'landingcard.dart';
